+++ data-raw/packages.R
@@ -40,7 +40,11 @@ paletteer_packages <- tibble::tribble(
   "vapeplot", "seasmith/vapeplot", "0.1.0", FALSE, NA,
   "ggthemr", "cttobin/ggthemr", "1.1.0", FALSE, NA,
   "fishualize", "nschiett/fishualize", "0.2.999", TRUE, "0.1.0",
-  "DresdenColor", "katiesaund/DresdenColor", "0.0.0.9000", FALSE, NA
+  "DresdenColor", "katiesaund/DresdenColor", "0.0.0.9000", FALSE, NA,
+  "trekcolors", "leonawicz/trekcolors", "0.1.2", TRUE, "0.1.1",
+  "nationalparkcolors", "katiejolly/nationalparkcolors", "0.1.0", FALSE, NA,
+  "MapPalettes", "disarm-platform/MapPalettes", "0.0.2", FALSE, NA,
+  "werpals", "sciencificity/werpals", "0.1.0", FALSE, NA
 ) %>%
   arrange(Name)